requirements.txt
python-jobspy>=1.1.0
pandas>=2.0.0
pyyaml>=6.0
